# Keep osd opwq worker wake for following op