--- a/src/common/config_opts.h
+++ b/src/common/config_opts.h
@@ -499,7 +499,7 @@ OPTION(osd_recovery_threads, OPT_INT, 1)
 OPTION(osd_recover_clone_overlap, OPT_BOOL, true)   // preserve clone_overlap during recovery/migration
 OPTION(osd_op_num_threads_per_shard, OPT_INT, 2)
 OPTION(osd_op_num_shards, OPT_INT, 5)
-OPTION(osd_op_worker_wake_time, OPT_INT, 10 * 1000) //us, zero for sleep at once
+OPTION(osd_op_worker_wake_time, OPT_INT, 0) //us, zero for sleep at once
 
 // Only use clone_overlap for recovery if there are fewer than
 // osd_recover_clone_overlap_limit entries in the overlap set