--- a/src/osd/OSD.cc
+++ b/src/osd/OSD.cc
@@ -8257,21 +8257,22 @@ void OSD::ShardedOpWQ::_process(uint32_t thread_index, heartbeat_handle_d *hb ) 
   ShardData* sdata = shard_list[shard_index];
   assert(NULL != sdata);
   sdata->sdata_op_ordering_lock.Lock();
-  
-  //keep wake for the next op for osd_op_worker_wake_time (default 10ms)
-  if (sdata->pqueue.empty() 
-      && sdata->has_waiting_thread == false 
+
+  //keep wake for the next op for osd_op_worker_wake_time if it is not zero
+  //By default (osd_op_worker_wake_time = 0), the worker will sleep immediately
+  if (sdata->pqueue.empty()
+      && sdata->has_waiting_thread == false
       && worker_thread_wake_time != (double)0) {
     sdata->has_waiting_thread = true;
     utime_t start_time = ceph_clock_now(osd->cct);
-    while (sdata->pqueue.empty()) {  
+    while (sdata->pqueue.empty()) {
       //give chance for enqueue
-      sdata->sdata_op_ordering_lock.Unlock(); 
-      
+      sdata->sdata_op_ordering_lock.Unlock();
+
       if (ceph_clock_now(osd->cct) - start_time >= worker_thread_wake_time) {
         //time out, go to sleep
         sdata->sdata_op_ordering_lock.Lock();
-        break; 
+        break;
       } else {
         //check the pqueue again
         sdata->sdata_op_ordering_lock.Lock();
@@ -8279,8 +8280,7 @@ void OSD::ShardedOpWQ::_process(uint32_t thread_index, heartbeat_handle_d *hb ) 
     }
     sdata->has_waiting_thread = false;
   }
-  
-  
+
   if (sdata->pqueue.empty()) {
     sdata->sdata_op_ordering_lock.Unlock();
     osd->cct->get_heartbeat_map()->reset_timeout(hb, 4, 0);
@@ -8363,7 +8363,7 @@ void OSD::ShardedOpWQ::_enqueue(pair<PGRef, OpRequestRef> item) {
   else
     sdata->pqueue.enqueue(item.second->get_req()->get_source_inst(),
       priority, cost, item);
-      
+
   if (sdata->has_waiting_thread) {
     sdata->sdata_op_ordering_lock.Unlock();
   } else {