@liewegas 2015-02-01
Closing this until we decide this is a good idea.  @yuandong1222 if you see this making a difference let us know!

I'm wondering if something a bit simpler would work just as well: drop and retake the lock (no delay) and see if there is another request that got queued.  Otherwise, go to sleep as per usual...