@yuandong1222 2015-02-02
Since now the msg use fast_dispatch to process MOSDOp, is it possible to
use msg worker thread (now it is pipe::reader and in the future it may be
AsyncMessenger::worker_thread) to run ReplicatedPG::do_request instead of
current OSD::OpWQ?

It is like Leader-Follower pattern.

On 1 February 2015 at 21:39, Sage Weil <notifications@github.com> wrote:

> Closed #2727 <https://github.com/ceph/ceph/pull/2727>.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/ceph/ceph/pull/2727#event-228131018>.
>



-- 
Dong Yuan
Email:yuandong1222@gmail.com